fix: Fix unlinking existing binaries (#12)
We failed to unlink binaries because we checked for the existence of
"path" instead of the correct variable name "outputPath".  This went
unnoticed because "path" is a global referring to the "path" module,
so the code still ran.

Another factor that made this more difficult to notice is that in
other methods, we clobbered the global module "path" with a method
parameter of the same name.  This has also been fixed by renaming
method parameters from "path" to something else.