fix: Fix Firefox macOS hang
Firefox version detection on macOS started hanging at some point.  Using osascript to get the version still works for other apps, but not Firefox, at least on my Intel mac.

This uses an alternative to get the version number on macOS, by parsing an ini file that Firefox seems to have as part of the .app bundle.

Closes #47