fix(ci): Trusted publishing works now (#57)
This was all tested and verified on my fork.

This also includes some fixes for publication from forks in general,
some of which intersect trusted publishing, and all of which I needed to
test this workflow properly from my fork.

To make OIDC (trusted publishing) work, we must:
 - set the required permission id-token: write (as documented)
 - update npm to 11.5.1 (as documented)
- update to node 24 (**not** documented, and **contradicting** examples
from npm)
- set matching repository.url in package.json (**not** documented, and
relevant to forks)

Some things that people suggested in forums, but which are not helpful:
- disabling "Require two-factor authentication for write actions" in
your account settings (no effect)
- selecting "Require two-factor authentication or a granular access
token with bypass 2fa enabled" in package settings (no effect)
 - passing --provenance to npm publish (no effect)